# api/openapi.yaml
openapi: 3.0.3
info:
  title: 今日のまとめアプリ
  version: 1.0.0
servers:
  - url: http://localhost:3000
tags:
  - name: diary
    description: 毎日の日記
  - name: todo
    description: 毎日のタスク
  - name: user
    description: ユーザー
paths:
  /diary:
    get:
      tags:
        - diary
      summary: 日記全件取得
      description: 日記内容の一部を全件取得します
      operationId: getDiary
      security:
        - Bearer: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Diarys"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /diary/create:
    post:
      tags:
        - diary
      summary: 日記作成
      description: 日記を新規作成します
      operationId: createDiary
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiaryOne"
        required: true
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /diary/getOne:
    get:
      tags:
        - diary
      summary: 日記一件取得
      description: idに合致する日記を一件取得します
      operationId: getDiaryById
      security:
        - Bearer: []
      parameters:
        - name: "diaryId"
          in: query
          description: 取得したい日記のID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryOne"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /diary/update:
    put:
      tags:
        - diary
      summary: 日記更新
      description: idに合致する日記を一件更新します
      operationId: updateDiary
      security:
        - Bearer: []
      parameters:
        - name: "diaryId"
          in: query
          description: 更新したい日記のID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDiaryOne"
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /diary/delete:
    delete:
      tags:
        - diary
      summary: 日記削除
      description: idに合致する日記を一件削除します
      operationId: deleteDiary
      security:
        - Bearer: []
      parameters:
        - name: "diaryId"
          in: query
          description: 更新したい日記のID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /todo:
    get:
      tags:
        - todo
      summary: タスク一覧取得
      description: タスク内容の一部を全件取得します
      operationId: getTodo
      security:
        - Bearer: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todos"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /todo/create:
    post:
      tags:
        - todo
      summary: タスク作成
      description: タスクを新規作成します
      operationId: createTodo
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoOne"
        required: true
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /todo/getOne:
    get:
      tags:
        - todo
      summary: タスク詳細取得
      description: idに合致するタスクを一件取得します
      operationId: getTodoById
      security:
        - Bearer: []
      parameters:
        - name: "todoId"
          in: query
          description: 取得したいタスクのID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoOne"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /todo/update:
    put:
      tags:
        - todo
      summary: タスク更新
      description: idに合致するタスクを一件更新します
      operationId: updateTodo
      security:
        - Bearer: []
      parameters:
        - name: "todoId"
          in: query
          description: 更新したいタスクのID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodo"
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /todo/delete:
    delete:
      tags:
        - todo
      summary: タスク削除
      description: idに合致するタスクを一件削除します
      operationId: deleteTodo
      security:
        - Bearer: []
      parameters:
        - name: "todoId"
          in: query
          description: 更新したいタスクのID
          required: true
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /auth/login:
    post:
      tags:
        - user
      summary: ユーザーログイン
      description: emailとパスワードを照合します
      operationId: loginUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUser"
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginUserResponse"
        "401":
          description: パスワードが間違っています
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /user/create:
    post:
      tags:
        - user
      summary: ユーザー作成
      description: ユーザーを新規作成します
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "409":
          description: メールアドレスがすでに登録されています
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /user/update:
    put:
      tags:
        - user
      summary: ユーザー更新
      description: idに合致するユーザーを一件更新します
      operationId: updateUser
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUser"
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: 認証されませんでした
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        "404":
          description: ユーザーが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /user/delete:
    delete:
      tags:
        - user
      summary: ユーザー削除
      description: idに合致するユーザーを一件削除します
      operationId: deleteUser
      security:
        - Bearer: []
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          description: 認証されませんでした
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedResponse"
        "404":
          description: ユーザーが見つかりません
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "500":
          description: サーバーエラー
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
components:
  securitySchemes:
      Bearer:
        type: http
        scheme: bearer
        description: Credentials or access token for API
  schemas:
    timeSchedule:
      type: array
      items:
        required:
          - StartTime
          - EndTime
          - todo
        type: object
        properties:
          StartTime:
            type: string
            format: 'date-time'
            example: '2024-10-07T14:30:00Z'
            description: '予定のの開始日時 (年-月-日T時:分:秒形式)'
          EndTime:
            type: string
            format: 'date-time'
            example: '2024-10-07T15:30:00Z'
            description: '予定のの終了日時 (年-月-日T時:分:秒形式)'
          todo:
            type: string
            example: 仕事
    Diarys:
      type: object
      properties:
        statusCode:
          type: integer
          description: ステータスコード
          example: 200
        data:
          type: array
          items:
            type: object
            required:
              - id
              - date
              - title
            properties:
              id:
                type: string
                format: uuid
                example: "123e4567-e89b-12d3-a456-426614174000"
              date:
                type: string
                format: date-time
                example: "2024-10-07T15:30:00Z"
              title:
                type: string
                example: "1日個人開発"
    CreateDiaryOne:
      required:
        - date
        - title
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2024-10-07T15:30:00Z"
        title:
          type: string
          example: 1日個人開発
        todo:
          type: array
          items:
            type: string
            format: uuid
        timeSchedule:
          $ref: "#/components/schemas/timeSchedule"
        timeScheduleResult:
          $ref: "#/components/schemas/timeSchedule"
        pros:
          type: array
          items:
            type: string
            example: 時間通り行動できた
        cons:
          type: array
          items:
            type: string
            example: 時間通り行動できなかった
        improvements:
          type: array
          items:
            type: string
            example: 時間通り行動するために時計を見る
        freeComment:
          type: string
          example: 今日は晴れだった。楽しかった。
    DiaryOne:
      type: object
      properties:
        statusCode:
          type: integer
          description: ステータスコード
          example: 200
        data:
          type: object
          required:
            - id
            - date
            - title
          properties:
            id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            date:
              type: string
              format: date-time
              example: "2024-10-07T15:30:00Z"
            title:
              type: string
              example: "1日個人開発"
            todo:
              type: array
              items:
                type: string
                format: uuid
            timeSchedule:
              $ref: "#/components/schemas/timeSchedule"
            timeScheduleResult:
              $ref: "#/components/schemas/timeSchedule"
            pros:
              type: array
              items:
                type: string
                example: "時間通り行動できた"
                description: "良かったところ"
            cons:
              type: array
              items:
                type: string
                example: "時間通り行動できなかった"
    Todos:
      type: object
      properties:
        statusCode:
          type: integer
          description: ステータスコード
          example: 200
        data:
          type: array
          items:
            required:
              - id
              - title
              - isFinished
            type: object
            properties:
              id:
                type: string
                format: uuid
                example: "123e4567-e89b-12d3-a456-426614174000"
              title:
                type: string
                example: 朝ごはんを食べる
              startDate:
                type: string
                format: 'date-time'
                example: '2024-10-07T15:30:00Z'
                description: 'タスクの開始日時 (年-月-日T時:分:秒形式)'
              timeRequired:
                type: integer
                example: 60
                description: 'タスクの所要時間'
              isFinished:
                type: boolean
                example: false
    CreateTodoOne:
      required:
        - title
      type: object
      properties:
        title:
          type: string
          example: 朝ごはんを食べる
        todoDetail:
          type: string
          example: りんごとパンを食べる
        startDate:
          type: string
          format: 'date-time'
          example: '2024-10-07T15:30:00Z'
          description: 'タスクの開始日時 (年-月-日T時:分:秒形式)'
        timeRequired:
            type: integer
            example: 60
            description: 'タスクの所要時間'
    TodoOne:
      type: object
      properties:
        statusCode:
          type: integer
          description: ステータスコード
          example: 200
        data:
          type: object
          required:
            - id
            - title
            - isFinished
          properties:
            id:
              type: string
              format: uuid
              example: "123e4567-e89b-12d3-a456-426614174000"
            title:
              type: string
              example: 朝ごはんを食べる
            todoDetail:
              type: string
              example: りんごとパンを食べる
            startDate:
              type: string
              format: 'date-time'
              example: '2024-10-07T15:30:00Z'
              description: 'タスクの開始日時 (年-月-日T時:分:秒形式)'
            timeRequired:
                type: integer
                example: 60
                description: 'タスクの所要時間'
            isFinished:
              type: boolean
              example: false
    UpdateTodo:
      required:
        - title
        - isFinished
      type: object
      properties:
        title:
          type: string
          example: 朝ごはんを食べる
        todoDetail:
          type: string
          example: りんごとパンを食べる
        startDate:
          type: string
          format: 'date-time'
          example: '2024-10-07T15:30:00Z'
          description: 'タスクの開始日時 (年-月-日T時:分:秒形式)'
        timeRequired:
            type: integer
            example: 60
            description: 'タスクの所要時間'
        isFinished:
          type: boolean
          example: false
    CreateUser:
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
          example: かぴ
        email:
          type: string
          example: Kapi@email.com
        password:
          type: string
          example: "12345abcde"
    LoginUser:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
          example: Kapi@email.com
        password:
          type: string
          example: "12345abcde"
    LoginUserResponse:
      required:
        - token
        - user
      type: object
      properties:
        token:
          type: string
          example: "aaaaa"
        user:
          type: object
          required:
            - username
            - useremail
          properties:
            username:
              type: string
              example: "maki"
            useremail:
              type: string
              example: "b@b"
    UpdateUser:
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
          example: かぴ
        email:
          type: string
          example: Kapi@email.com
        password:
          type: string
          example:  "12345abcde"
    Response:
      type: object
      properties:
        message:
          type: string
          description: メッセージ
        statuscode:
          type: integer
          description: ステータスコード
    UnauthorizedResponse:
      type: object
      properties:
        message:
          type: string
          description: エラーメッセージ
          example: "Unauthorized"
        statusCode:
          type: integer
          description: HTTP ステータスコード
          example: 401